Fix XLSX helper import cycle
